feat(ds): Add story data support
Attach data to a story config or a single story via the new `data` key available in the
story configuration schema. It may be path string or inline object. The data is available
as `config.data` respectively `story.data` within the nunjucks story templates. Demonstration of its usage is available
within the newly introduced `card` molecule (and in the test snapshots).

Another notable changes in this PR:
 - update to the latest alpha and build scripts
 - refactor the modules to control the styles
 - refactor test app folder
 - simplify the code tag usage (breaking change)

closes #17